refactor: []byte <-> string, named types
this commit is an attempt to bring in some consistency of data types and reduce random inter-conversions across the whole project. firstly, wherever possible, we try to keep []byte over string. usage of string is needed in two places: when user has to view the data and when key needs to be stored in map since []byte can't be hashed. so we keep []byte across the system and convert to string only for the aforementioned cases. besides this, we wrap this up using a named type which enables ease of change and reduces cognitive complexity.