chore: move stdout away from library
it is better if stdout operations are not spread out all over the place
stdout is for humans and the library does not interact with humans. cli
does. the libraries are better off with just passing messages and
letting the interaction layer decide how, when and if the output should
be displayed.